+++ Client RLD/src/Tile.cpp
@@ -7,6 +7,6 @@ Tile::Tile(TileType type, sf::Texture& tex) {
 
 void Tile::draw(sf::RenderWindow& window, int x, int y) {
 	//Logger::log(std::to_string(x) +  std::to_string(y));
-	this->sprite.setPosition(x, y);
+	this->sprite.setPosition((float)x, (float)y);
 	window.draw(this->sprite);
 };
